fix: Resolving Mold Differences

+++ 2021_RTCSA_dynfed/src/dynfed.rs
@@ -6,10 +6,10 @@ use std::collections::HashMap;
 
 #[allow(dead_code)] // TODO: remove
 pub fn pre_computation(
-    dag: &mut Graph<NodeData, f32>,
+    dag: &mut Graph<NodeData, i32>,
 ) -> HashMap<Graph<NodeData, i32>, Vec<NodeIndex>> {
-    let vol = dag.get_volume() as i32;
-    let end_to_end_deadline = dag.get_end_to_end_deadline().unwrap() as i32;
+    let vol = dag.get_volume();
+    let end_to_end_deadline = dag.get_end_to_end_deadline().unwrap();
     let m_min = (vol as f32 / end_to_end_deadline as f32).ceil() as usize;
     let execution_order: HashMap<Graph<NodeData, i32>, Vec<NodeIndex>> = HashMap::new();
